requirements.txt
sqlalchemy>=2.0.32
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
google-api-python-client>=2.140.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
pytest>=8.3.2


